Creates PoC pipeline to estimate disk usage of vcf_to_bq on Dataflow.
The pipeline uses raw file size and raw+encoded sizes of a short snippet at beginning of VCF files to
estimate the encoded size for a commit. The major blocking bug is that when the snippets are being read from VCFs in an
encoded format, lines are being read more than once.

intermediate commit